[refactor/#274] notificationDataStore로 분리
