+++ data/local/src/main/java/com/napzak/market/local/datastore/TokenDataStore.kt
@@ -2,9 +2,9 @@ package com.napzak.market.local.datastore
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
+import com.napzak.market.local.datastore.NotificationDataStore.Companion.preferencesNotificationModalShown
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -20,18 +20,6 @@ class TokenDataStore @Inject constructor(
         .map { it[preferencesRefreshTokenKey] }
         .firstOrNull()
 
-    suspend fun getPushToken(): String? = preferenceDataStore.data
-        .map { it[preferencesPushTokenKey] }
-        .firstOrNull()
-
-    suspend fun getNotificationPermission(): Boolean? = preferenceDataStore.data
-        .map { it[preferencesNotificationPermission] }
-        .firstOrNull()
-
-    suspend fun getNotificationModalShown(): Boolean? = preferenceDataStore.data
-        .map { it[preferencesNotificationModalShown] }
-        .firstOrNull()
-
     suspend fun setTokens(accessToken: String, refreshToken: String) {
         preferenceDataStore.edit { preferences ->
             preferences[preferencesAccessTokenKey] = accessToken
@@ -39,28 +27,10 @@ class TokenDataStore @Inject constructor(
         }
     }
 
-    suspend fun setPushToken(pushToken: String) {
-        preferenceDataStore.edit { preferences ->
-            preferences[preferencesPushTokenKey] = pushToken
-        }
-    }
-
-    suspend fun setNotificationPermission(permission: Boolean) {
-        preferenceDataStore.edit { preferences ->
-            preferences[preferencesNotificationPermission] = permission
-        }
-    }
-
     suspend fun clearTokens() {
         preferenceDataStore.edit { preferences ->
             preferences.remove(preferencesAccessTokenKey)
             preferences.remove(preferencesRefreshTokenKey)
-        }
-    }
-
-    suspend fun clearPushToken() {
-        preferenceDataStore.edit { preferences ->
-            preferences.remove(preferencesPushTokenKey)
         }
     }
 
@@ -81,23 +51,11 @@ class TokenDataStore @Inject constructor(
         }
     }
 
-    suspend fun updateNotificationModalShown() {
-        preferenceDataStore.edit { preferences ->
-            preferences[preferencesNotificationModalShown] = true
-        }
-    }
-
     companion object {
         private const val ACCESS_TOKEN_KEY = "access_token"
         private const val REFRESH_TOKEN_KEY = "refresh_token"
-        private const val PUSH_TOKEN_KEY = "push_token"
-        private const val NOTIFICATION_PERMISSION_KEY = "notification_permission"
-        private const val NOTIFICATION_MODAL_SHOWN_KEY = "notification_modal_shown"
 
         val preferencesAccessTokenKey = stringPreferencesKey(ACCESS_TOKEN_KEY)
         val preferencesRefreshTokenKey = stringPreferencesKey(REFRESH_TOKEN_KEY)
-        val preferencesPushTokenKey = stringPreferencesKey(PUSH_TOKEN_KEY)
-        val preferencesNotificationPermission = booleanPreferencesKey(NOTIFICATION_PERMISSION_KEY)
-        val preferencesNotificationModalShown = booleanPreferencesKey(NOTIFICATION_MODAL_SHOWN_KEY)
     }
 }
